+++ tests/AppTest.php
@@ -12,7 +12,7 @@ class AppTest extends TestCase
   /**
    * Setup test
    */
-  public function setup()
+  public function setUp(): void
   {
     global $app;