+++ tests/Core/ContainerTest.php
@@ -25,7 +25,7 @@ class ContainerTest extends TestCase
   /**
    * Setup test
    */
-  public function setup()
+  public function setUp(): void
   {
   }